Rename the chrome module to back
This repo embeds Chromium, writes a `chrome_debug.log`, and already spells the
UI sense of the word only in qualified form (`ExplorerChrome`). A bare `chrome`
module sitting in `vmux_ui` reads as the browser engine at a glance and has to
be opened to find out it is not.

`back` says what is in it — `PageBack` and the button that calls it — and does
not stutter at the use site. The docs no longer lean on the word either: what
they are describing is a host that surrounds a page with nothing of its own,
which is plainer said that way.
